smol-marketing-claw/requirements.txt
mcp>=1.0.0
uvicorn>=0.27.0
aiohttp==3.10.0
discord.py==2.3.2
python-dotenv==1.0.0
tweepy==4.14.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
